add support for pre-shared keys
Allow the client to generate and pass a PSK to the server, which is then
applied to the respective WireGuard peer.

This is useful because unlike regular WireGuard, TLS connections can now
support post-quantum key exchanges[1] and if we exchange a PSK through
such a TLS connection, we inherit its PQ-safety.

[1] https://datatracker.ietf.org/doc/draft-ietf-tls-ecdhe-mlkem/